move widget into own screens folders
