+++ lib/screens/accountScreens/widgets/reviewCard.dart
@@ -2,8 +2,8 @@ import 'package:Pitch/models/reviewList.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import '../appTheme.dart';
-import 'starRating.dart';
+import '../../../appTheme.dart';
+import '../../../widgets/starRating.dart';
 
 class ReviewCard extends StatefulWidget {
   final AnimationController animationController;
