+++ lib/screens/favoriteScreens/widgets/favoriteCampingCard.dart
@@ -3,7 +3,7 @@ import 'package:Pitch/widgets/starRating.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-import '../appTheme.dart';
+import '../../../appTheme.dart';
 
 class FavoriteCampingCard extends StatefulWidget {
   final AnimationController animationController;
