+++ lib/screens/homepage/homeScreen.dart
@@ -1,11 +1,11 @@
 import 'package:Pitch/screens/settingsScreens/settingScreen.dart';
-import 'package:Pitch/widgets/FilterPremium.dart';
-import 'package:Pitch/widgets/campingList.dart';
+import 'package:Pitch/screens/homepage/widget/FilterPremium.dart';
+import 'package:Pitch/screens/homepage/widget/campingList.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import '../../appTheme.dart';
 import '../accountScreens/accountScreen.dart';
-import '../favoritesScreen.dart';
+import '../favoriteScreens/favoritesScreen.dart';
 import '../searchScreens/searchScreen.dart';
 
 class HomeScreen extends StatefulWidget {
@@ -23,12 +23,10 @@ class _HomeScreen extends State<HomeScreen>
 
   int _currentIndex = 0;
   final List<Widget> _children = [
-    //HomeScreen(),
     Center(child: Text('Home in lavorazione')), //home body
-    FavoritesScreen(), //Center(child: Text('Preferiti in lavorazione')),  //favorite body
-    //SearchPage(), //al suo posto la nuova funzionalità
+    FavoritesScreen(), 
+    // nuova funzionalità
     AccountScreen(),
-    //Center(child: Text('Profilo in lavorazione')),  //account body
   ];
 
   void initState() {
@@ -55,46 +53,6 @@ class _HomeScreen extends State<HomeScreen>
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.grey[100],
-      appBar: _currentIndex == 0
-          ? null
-          // ? AppBar(
-          //     title: Row(
-          //       mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          //       children: <Widget>[
-          //         Row(
-          //           children: <Widget>[
-          //             //logo image
-          //             Padding(
-          //                 padding: const EdgeInsets.only(left: 8.0, top: 8.0),
-          //                 child: Image.asset(
-          //                   'assets/Icons/logo.png',
-          //                   fit: BoxFit.cover,
-          //                   height: 40,
-          //                 )),
-          //           ],
-          //         ),
-          //       ],
-          //     ),
-          //     backgroundColor: Colors.white,
-          //     actions: <Widget>[
-          //       IconButton(
-          //         icon: const Icon(Icons.settings),
-          //         tooltip: 'Show Snackbar',
-          //         onPressed: () {
-          //           Navigator.push(
-          //             context,
-          //             MaterialPageRoute(
-          //                 builder: (context) => SettingScreen(),
-          //                 fullscreenDialog: true),
-          //           );
-          //         },
-          //         color: Colors.black,
-          //       ),
-          //     ],
-          //   )
-          : AppBar(
-              toolbarHeight: 0,
-            ),
       body: _currentIndex == 0 ? getHomePage() : _children[_currentIndex],
       // ignore: missing_required_param
       bottomNavigationBar: BottomNavigationBar(
