+++ lib/screens/searchScreens/searchScreen.dart
@@ -2,6 +2,7 @@ import 'package:Pitch/appTheme.dart';
 import 'package:Pitch/models/campingList.dart';
 import 'package:Pitch/screens/popups/orderPopup.dart';
 import 'package:Pitch/widgets/campingCard.dart';
+import 'package:Pitch/widgets/customAppBar.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:intl/intl.dart';
